requirements.txt
requests==2.31.0
google-generativeai==0.3.2
python-dotenv==1.0.0
schedule==1.2.0
Pillow==10.0.1
croniter==1.4.1
Flask==2.3.3